requirements-dev.txt
# ===============================================================================
# REQUIREMENTS PARA DESARROLLO LOCAL EN WINDOWS
# ===============================================================================
# 
# Este archivo es ÚNICAMENTE para desarrollo local en Windows con problemas
# de compatibilidad (Python 3.13 + ARM64, falta de herramientas de compilación)
# 
# PARA CUALQUIER OTRO ENTORNO USA: requirements.txt
# - Docker (local o en la nube)  
# - Linux/macOS
# - Servidores de producción
# - CI/CD pipelines
# 
# ===============================================================================

# === DIFERENCIAS PRINCIPALES CON requirements.txt ===
# 1. pg8000 en lugar de psycopg2-binary (no compila en Windows ARM64)
# 2. uvicorn sin [standard] (httptools no compila)  
# 3. Versiones actualizadas para compatibilidad con Python 3.13

# === DEPENDENCIAS CORE ===
fastapi==0.104.1
uvicorn==0.24.0  # Sin [standard] para evitar problemas de compilación

# Dependencias manuales de uvicorn
click==8.1.7
h11==0.14.0

# === BASE DE DATOS ===
sqlalchemy==2.0.43  # Versión actualizada compatible con Python 3.13
pg8000==1.31.5      # Reemplaza psycopg2-binary
alembic==1.13.0

# === VALIDACIÓN Y SERIALIZACIÓN ===
pydantic>=2.10.0  # Versión más reciente compatible con Python 3.13
pydantic[email]>=2.10.0
email-validator==2.1.0

# === AUTENTICACIÓN Y SEGURIDAD ===
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# === UTILIDADES ===
python-dotenv==1.0.0
python-multipart==0.0.6

# === LOGGING ===
structlog==23.2.0
python-json-logger==2.0.7

# === TESTING ===
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# === MONITORING ===
prometheus-client==0.19.0

# === MESSAGING ===
pika==1.3.2
celery==5.3.4

# ===============================================================================
# INSTALACIÓN:
# pip install -r requirements-dev.txt
# 
# CONFIGURACIÓN DE BD:
# DATABASE_URL=postgresql+pg8000://postgres:password@localhost:5432/donations_db
# ===============================================================================